only BatchExecutor buffers the transformations, and make sure that the explain method is stateless
